feature: 투두 삭제기능 추가 - query, command 분리 (controller, service)
