+++ src/main/java/com/godlife/goalservice/repository/TodoTaskScheduleRepository.java
@@ -1,9 +1,11 @@
 package com.godlife.goalservice.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.godlife.goalservice.domain.TodoTaskSchedule;
 
 public interface TodoTaskScheduleRepository extends JpaRepository<TodoTaskSchedule, Long>, TodoTaskScheduleRepositoryCustom {
-
+	List<TodoTaskSchedule> findAllByTodoTaskTodoId(Long todoId);
 }
