--- src/test/java/com/godlife/goalservice/GoalServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.godlife.goalservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class GoalServiceApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
